Get rid of unnecessary magic ["table"] RecursiveOpenStruct access
This incantation was a workaround for recursive-open-struct < 1.0.3
bug preventing even [:namespace] access to a field, when a #namespace
method existed.
But it was a weird way to access the hash, relying on a bug that's fixed
in same 1.0.3.
[https://github.com/aetherknight/recursive-open-struct/issues/51]

The Object#namespace monkey patch was removed in Aug 2015
[https://github.com/ManageIQ/more_core_extensions/pull/20]
so we can just use `.namespace` access.
(works with both old and new recursive-open-struct).