Lazy load kubeclient/image inspector library with autoload
Note, there doesn't seem to be one single entrypoint for these classes and
mixins so when possible, we now autoload any constants being used in these
files.

Note, it's best to require or use ruby's autoload on libraries like this.
Normal rails models and other files that need to be reloaded by rails in dev
mode need to be loaded by using the constant directly to force const_missing
or use require_dependency.