Use an inventory_collection to raise created events
Instead of an extra method in the Refresher class we can raise creation
events with an inventory_collection which is more consistent with how
this is typically done and we can clean up the base Refresher class.

Containers were the only providers that actually used this
manager_refresh_post_processing method and it can be deleted now.