WIP Add some debugging for kubeclient
Kubeclient uses RestClient for 4.x versions for the main connection
but the http gem for watches.  On master, kubeclient uses faraday for
main connections.  This commit provides a way to patch restclient/faraday
and http to add logging of request/response.

TODO: We will need to provide a way to sanitize passwords and possibly other
information.