Fix prometheus http_proxy using 0.0.0.0:80
When `VMDB::Util.http_proxy_uri` is `nil`
`VMDB::Util.http_proxy_uri.to_s` is `""`.

With uri v0.10.3 `URI.parse("http://")` would have a `host` of `nil` but
on v0.13.0 `URI.parse("http://")` has a `host` of `""`.

This was causing Faraday to use a proxy of `0.0.0.0:80` instead of none.